Fixed flat render depth and added a test crosshair

+++ src/main/java/ru/windcorp/optica/client/graphics/Layer.java
@@ -46,11 +46,11 @@ public abstract class Layer {
 
 	protected abstract void doRender();
 	
-	protected float getWidth() {
+	protected int getWidth() {
 		return GraphicsInterface.getFramebufferWidth();
 	}
 	
-	protected float getHeight() {
+	protected int getHeight() {
 		return GraphicsInterface.getFramebufferHeight();
 	}
